+++ lib/libwebsocket/handshake/server.rb
@@ -1,12 +1,12 @@
 module LibWebSocket
-  class OpeningHandshake
+  class Handshake
     # Construct or parse a server WebSocket handshake. This module is written for
     # convenience, since using request and response directly requires the same code
     # again and again.
     #
     # SYNOPSIS
     #
-    #   h = LibWebSocket::OpeningHandshake::Server.new
+    #   h = LibWebSocket::Handshake::Server.new
     #
     #   # Parse client request
     #   h.parse \<<EOF
@@ -32,7 +32,7 @@ module LibWebSocket
     #          # Sec-WebSocket-Location: ws://example.com/demo
     #          #
     #          # fQJ,fN/4F4!~K~MH
-    class Server < OpeningHandshake
+    class Server < Handshake
 
       # Parse client request
       # @example